autoblogger-legacy/requirements.txt
# AutoBlogger Dependencies - Latest Versions (2024)
# Core functionality
google-generativeai>=0.8.0
openai>=1.0.0
httpx>=0.27.0
pydantic>=2.10.0
python-dotenv>=1.0.1
pillow>=10.4.0
schedule>=1.2.2

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
pytest-cov>=6.0.0

# Code quality
black>=24.10.0
mypy>=1.13.0
flake8>=7.1.0

# Web interface
flask>=3.1.0
flask-cors>=5.0.0
werkzeug>=3.1.0
jinja2>=3.1.4

# Development
pre-commit>=4.0.0
